Keep shared/gems around to continue to benefit from shared code
The original lstoll heroku-buildpack-monorepo moved code around, but
that made it impossible to keep a directory such as shared/gems around.
Heroku's https://github.com/heroku/heroku-buildpack-multi-procfile
does support keeping shared/gems around, but breaks the Heroku model
because the app's root is not the root anymore and the heroku/ruby
buildpack does not know to build the app deeper in /app.

The changes in this commit improve the situation by moving shared/gems
around and editing (yuck!) Gemfile and Gemfile.lock to reference the new
location of shared/gems. This is hackish solution and I'm trying to find
a better way.